订单详情完成

+++ src/main/java/com/mapei/www/entity/Products.java
@@ -10,6 +10,6 @@ public class Products {
     private String image;
     private String productid;
     private Double listprice;
-    private Integer unitcost;
+    private Double unitcost;
     private Integer quantity;
 }
